+++ docs/PUBLIC_API.md
@@ -35,7 +35,7 @@ We explicitly don't consider the following to be our public API surface:
 - any class members or symbols marked as `private`, or prefixed with underscore (`_`), [barred latin o](https://en.wikipedia.org/wiki/%C6%9F) (`ɵ`), and  double barred latin o (`ɵɵ`).
 - extending any of our classes unless the support for this is specifically documented in the API docs
 - the contents and API surface of the code generated by Angular's compiler (with one notable exception: the existence and name of `NgModuleFactory` instances exported from generated code is guaranteed)
-
+- the `@angular/core/primitives` package, including its child entrypoints
 
 Our peer dependencies (such as TypeScript, Zone.js, or RxJS) are not considered part of our API surface, but they are included in our SemVer policies. We might update the required version of any of these dependencies in minor releases if the update doesn't cause breaking changes for Angular applications. Peer dependency updates that result in non-trivial breaking changes must be deferred to major Angular releases.
