+++ packages.bzl
@@ -63,7 +63,6 @@ DOCS_ENTRYPOINTS = [
     "core/src/di/interface",
     "core/src/interface",
     "core/src/reflection",
-    "core/src/signals",
     "core/src/util",
     "core/testing",
     "elements",
