+++ packages/bazel/test/ng_package/core_package.spec.ts
@@ -65,6 +65,12 @@ describe('@angular/core ng_package', () => {
               esm: './esm2022/core.mjs',
               default: './fesm2022/core.mjs'
             },
+            './primitives/signals': {
+              types: './primitives/signals/index.d.ts',
+              esm2022: './esm2022/primitives/signals/index.mjs',
+              esm: './esm2022/primitives/signals/index.mjs',
+              default: './fesm2022/primitives/signals.mjs',
+            },
             './rxjs-interop': {
               types: './rxjs-interop/index.d.ts',
               esm2022: './esm2022/rxjs-interop/rxjs-interop.mjs',
