+++ packages/core/primitives/signals/src/equality.ts
