+++ packages/core/primitives/signals/src/signal.ts
