+++ packages/core/test/acceptance/BUILD.bazel
@@ -23,7 +23,6 @@ ts_library(
         "//packages/common/locales",
         "//packages/compiler",
         "//packages/core",
-        "//packages/core/src/signals",
         "//packages/core/src/util",
         "//packages/core/test/render3:matchers",
         "//packages/core/testing",
